+++ src/cargo/util/vcs.rs
@@ -5,6 +5,21 @@ use git2;
 
 use util::{process, CargoResult};
 
+// Check if we are in an existing repo. We define that to be true if either:
+//
+// 1. We are in a git repo and the path to the new project is not an ignored
+//    path in that repo.
+// 2. We are in an HG repo.
+pub fn existing_vcs_repo(path: &Path, cwd: &Path) -> bool {
+    fn in_git_repo(path: &Path, cwd: &Path) -> bool {
+        if let Ok(repo) = GitRepo::discover(path, cwd) {
+            repo.is_path_ignored(path).map(|ignored| !ignored).unwrap_or(true)
+        } else { false }
+    }
+
+    in_git_repo(path, cwd) || HgRepo::discover(path, cwd).is_ok()
+}
+
 pub struct HgRepo;
 pub struct GitRepo;
 pub struct PijulRepo;
